Linux 2.2.17pre15
o	Pull the 3ware update - causes oopses ?		(Alan Cox)
o	Fix emu10k build error with no PCI		(Arjan van de Ven)
o	display7 LED driver 				(Eric Brower)
o	TLB multi-match bugs in sparc64			(Dave Miller)
o	Fix CPP string concatenation in sparc ksyms	(Dave Miller)
o	Fix lost copyright credit to ALi in the
	trident.c driver				(Alan Cox)
o	Add alternate ALi ID to the trident driver	(Alan Cox)
o	ATI video fixes for PPC				(Benjamin Herrenschmidt)
o	Further HDLC fixes				(Krzysztof Halasa)
o	Fix sign handling bugs				(Andrew Morton, Alan Cox)
o	Fix wrong comment in fs/exec.c			(Robert Love)

+++ drivers/char/epca.c
@@ -3833,7 +3833,7 @@ void epca_setup(char *str, int *ints)
 
 			case 5:
 				board.port = (unsigned char *)ints[index];
-				if (board.port <= 0)
+				if ((signed long)board.port <= 0)
 				{
 					printk(KERN_ERR "<Error> - epca_setup: Invalid io port 0x%x\n", (unsigned int)board.port);
 					invalid_lilo_config = 1;
@@ -3845,7 +3845,7 @@ void epca_setup(char *str, int *ints)
 
 			case 6:
 				board.membase = (unsigned char *)ints[index];
-				if (board.membase <= 0)
+				if ((signed long)board.membase <= 0)
 				{
 					printk(KERN_ERR "<Error> - epca_setup: Invalid memory base 0x%x\n",(unsigned int)board.membase);
 					invalid_lilo_config = 1;
